move api code to package that makes sense
